requirements.txt
azure-storage-blob
python-dotenv
notebook
jupytext
pandas
numpy
requests
tqdm
seaborn
click
scikit-learn